Code Improvement: Less lines & rebase
I have been reviewing the code and realized that there are a lot of
useless lines that doesnt gives anything to the project, so I decided to
remove them, I guess those lines wont make errors later.
Changed a little thing in the random formula, because someone explained
me why that formula wouldnt work

Rebase details:
I had to make changes from the first commit because there were lines
with IDs and more things that could compromise my personal information,
so I decided to make a new ignored folder with all the IDs and things,
and only the hosting page can acces to it with a different filter file
".gitIgnore". Also as Im new with rebase command I will have to commit
again everything and the push date will be July 7, but inside all the
commits there is a line with the real Upload date