New random rotor generator
I have been working on a order randomizer that creates random rotors depending on a seed, so It's not totally random.
To make these generator I used a cool library called random-seed, Its truly useful.
Also improved some things on the random functions
My plans for the future is to update the bot to v13, and stop using v12, Im not saying v12 is bad but I just want to learn and try slash commands and new cool features in v13
Next commit will be probably making the pin system or filter on the output or anything that comes to my mind.