New stuff
I forgot what I did in the last commit, it was a long ago. So, now you can have a lot of different characters and number inside the message and inside the configuration, there are a lot of random functions added in order to make a new seeded random generator for accessing some previus rotor configuration. Fixed a lot of stuff that was glitching, it was all because the amount of characters in the alphabet where odd and had to be even for the reflectors to work and have a good pair relation. Planning on working in the pin system and more stuff ig